\name{joinEasy}
\alias{joinEasy}
\docType{package}
\title{
\packageTitle{joinEasy}
}
\description{
\packageDescription{joinEasy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{joinEasy}
\packageIndices{joinEasy}
    \code{\link{anti_join}}
    \code{\link{cross_join}}
}
\author{
\packageAuthor{joinEasy}

Maintainer: \packageMaintainer{joinEasy}
}
\references{
www.mtitek.com/tutorials/oracle/sql-join.php Provides visual depictions of all operations present in this package
}
\keyword{ package }
\seealso{
#~~ Optional links to other man pages, e.g. ~~
#~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
